Pass discount codes to the SignupProcessor
This is so that we can pass the discount code onto the `Invoicer` so it
eventually ends up on the invoice.

Coupons are now registered (from Chargify) with their value so that we
can look it up again based on the Member's stored coupon code when they
are queued up in the SignupProcessor.