Fix problem where members are left in limbo
This happens if the member gets to the payment page and then realises
they cannot pay with a credit card, so they contact us to pay by invoice
but the member already exists so the process fails.

This just forwards them onto the thanks page so that the rest of the
process can continue.