Retrieve Chargify product links from ENV
This is a pretty big change.

At application startup, product signup page URLs were requested from
Chargify via their API and stored in a constant. These were then used to
forward members onto Chargify payment pages.

This technique is still used for other parts of the system that use
Chargify. Hopefully, we can do something about this in the future.

We needed to use a different payment page for the same product dependant
on a couple of factors.

Unfortunately, there is no identifier in the response from the Chargify
API that I can use in order to map products to signup pages, so I have
no alternative but to replace the products list with a hardcoded
implementation that comes from the environment.

We now just hardcode the values in environment variables,
`CHARGIFY_PRODUCT_URL` and `CHARGIFY_PAGE_IDS`

There isn't really any advantage of getting these from Chargify
automatically anyway. Even if we add Chargify products links
automatically, we still need to make code changes AND restart the app in
order to use them, so fetching them automatically only adds complexity
with little benefit.