Make all students use the same product internally
Before we were using `student-free` as the product name but this would
mean other changes to the system because the product name trickles
through.

This way we can keep all students on the same product but pass them to
different payment pages if `?free=true` is added to the signup page.